Migrating to goomba function
